Move dead code out of the way.
